+++ changelog.md
@@ -18,6 +18,7 @@
   - Label for "Screen Controls Position" now consistent with the styles of the rest
 - Internal:
   - Remove `radio-selection` component and general cleanup
+  - Update authentication to read `lastGame` from `player` property (API 0.3.2 change)
 - Add Valentine's animation
 
 # 0.8.7 Beta - 02/02/2026 13:30 UTC+1